Allow multiple calls to project methods without reinitializing project
Prior to this change if one was to do project.Create() and then
project.Start(), with the same project object, the second invocation
would do nothing.  This is because the initial list of services to
traverse was based off of p.reload.  This patch changes it such that
reload is not taken into account on start but only when we need to
restart the execution.

Signed-off-by: Darren Shepherd <darren@rancher.com>