Adding standard LoggerFactory for image building and pulling
This change updates the builder and puller to use the context.LoggerFactory
for image building, over the default, stdout.

By making this change, image build logs can be rendered in custom formats,
using the LoggerFactory provided in the docker context. This also
allows logs to be redirected to sinks other than stdout.

ColorLogger has been updated and a RawLogger type has been added to allow
for matching current behaviour.

Signed-off-by: Brendan Fosberry <brendan.fosberry@gmail.com>