Adding Failure events to event types and to service wrapper
This change adds a set of project and service level failure events to the
list of possible events, as well as to the service wrapper function calls.

This allows users to determine whether an action has failed by viewing events.

The only behaviour changed here, other than additional events passing through
the notification system, is that when a project action fails, it will no longer
post the relevant 'done' event.

Signed-off-by: Brendan Fosberry <brendan.fosberry@gmail.com>