Remove logrus dependency in the core of the library
A library should not use a specific logger over an other (logrus
here), but let the users of the library choose theirs. If it really
needs to log things, it should provide a composable way to do it.

This refactoring defines a new type Logger and rename the old one
into LogFormatter. A default Logger (based on fmt) is defined and used
if nothing is used and `logrus.New()` can be used directly as
logger (and that's what used for the cli)

Signed-off-by: Vincent Demeester <vincent@sbr.pm>