Use context.Context accross the library
Applications that require full plumbing of context, such as those that
rely on cancellation for resource cleanup, would be incompatible with
inconsistent plumbing.

This *connects* the net/context from the user of the libcompose API to
the engine-api and below (where request are cancellable for
example). This should also allow us to some cool and safe stuff with
concurrent goroutines and cancellation O:)

Signed-off-by: Vincent Demeester <vincent@sbr.pm>