Introduce a RuntimeProject interface
The `RuntimeProject` interface defines project-level methods that need
direct access to the *backend* runtime (i.e. docker) 🐮.

This allows to not depend on docker in `project` package and thus,
making it less coupled to any runtime.

This also moves back `ClientFactory` to `docker/client` package as it is
specific to the docker runtime *and* is only required in `docker`
package.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>