Parse multifile COMPOSE_FILE envvar, w/ side effect
Fixes #212

urfave/cli does not distinguish whether the first string in the slice
comes from the envvar or is from a flag. Worse off, it appends the flag
values to the envvar value instead of overriding it. To ensure the
multifile envvar case is always handled, the first string must always be
split. It gives a more consistent behavior, then, to split each string
in the slice. With this side effect, the following is possible and
would merge the compose files left to right:

```
$ COMPOSE_FILE=a.yml:b.yml ./libcompose-cli -f c.yml -f d.yml:e.yml up
```

Signed-off-by: Dustin Chapman <dustin.r.chapman@gmail.com>