Create new image and auth packages
The image package is responsible to handle operation on images (inspect,
pull, push, …).

The auth package defines a way to lookup for Auth information (way to
log in to registries) and provide an implementation based on the
configfile implement in docker/docker.

Both package might need refactoring later, it's just an extraction of
the `image.go` and `auth.go` files.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>