Add a config command
Add a simple config command (without any options for now) that does the
same as `docker-compose config`.

It's pretty early on and the output will need tweaking. Even with
tweaking it is going to be difficult to get the exact same output as
docker-compose. The main point is to produce valid yaml that can be
reused by both of them.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>