Add ServiceNames() to APIProject
Issue Observed
--------------

project.APIProject doesn't provide a method to get the service names
define in a project.

You cannot get service names when you use GetSreviceConfig() and other
methods of the interface.

Proposed Resolution
-------------------

We could add a method GetServiceConfigs() which returns the
*ServiceConfigs owned by the concrete struct (project.Project).
However, since the strcut already has ServiceConfigs property, we cannot add a new method with the same name.

In this commit, we add a new method just returns the list of service
names instead of returning the *ServiceConfigs itself.

Signed-off-by: Iwasaki Yudai <yudai@arielworks.com>